Update: ProgressViewStyle 커스텀 - 승자, 패자 분기를 나눌 수 있는 isWin 파라미터 추가 - PercentageBar -> CustomProgressView로 이름 변경 - 필요없는 개행 삭제
